+++ HotelApp/Controllers/RoomController.cs
@@ -15,6 +15,8 @@ namespace HotelApp.API.Controllers
     using Model.Room;
     using Model.Hotel;
     using AutoMapper.QueryableExtensions;
+    using Microsoft.Extensions.Caching.Memory;
+    using System.Threading;
 
     [Route("api/hotel/{hotelId}/room")]
     [ApiController]
@@ -23,19 +25,21 @@ namespace HotelApp.API.Controllers
         private readonly ApplicationDbContext _context;
         private readonly ILogger<RoomController> _logger;
         private readonly IMapper _mapper;
+        private readonly IMemoryCache _memoryCache;
 
-        public RoomController(ApplicationDbContext context, ILogger<RoomController> logger, IMapper mapper)
+        public RoomController(ApplicationDbContext context, ILogger<RoomController> logger, IMapper mapper, IMemoryCache memoryCache)
         {
             _context = context;
             _logger = logger;
             _mapper = mapper;
+            _memoryCache = memoryCache;
         }
 
         [HttpPost]
-        public async Task<ActionResult<Room>> Post(int hotelId, CreateRoomResource model)
+        public async Task<ActionResult<Room>> Post(int hotelId, CreateRoomResource model, CancellationToken token)
         {
             //map model to entity
-            var hotel = await this._context.Hotels.FindAsync(hotelId);
+            var hotel = await this._context.Hotels.FindAsync(new object[] { hotelId }, token);
 
             if (hotel == null)
             {
@@ -44,36 +48,58 @@ namespace HotelApp.API.Controllers
 
             var entity = this._mapper.Map<Room>(model);
             entity.Hotel = hotel;
+
+            if (token.IsCancellationRequested) return this.NoContent();
+
             this._context.Rooms.Add(entity);
 
-            await this._context.SaveChangesAsync();
+            await this._context.SaveChangesAsync(token);
+
+            this._memoryCache.Remove($"_HOTEL_ROOMS_{hotelId}");
 
             return this.CreatedAtAction("Get", new { hotelId, id = entity.Id }, this._mapper.Map<RoomResource>(entity));
         }
 
         [HttpGet("")]
-        public async Task<ActionResult<IEnumerable<RoomResource>>> GetRooms(int hotelId)
+        public async Task<ActionResult<IEnumerable<RoomResource>>> GetRooms(int hotelId, CancellationToken token)
         {
             this._logger.LogInformation("RoomController-GetRooms hit");
 
-            var list = await this._context.Rooms
-                .Include(h => h.Hotel)
-                .Where(h => h.Hotel.Id == hotelId)
-                .ToListAsync();
+            var key = $"_HOTEL_ROOMS_{hotelId}";
+            if (_memoryCache.TryGetValue(key, out List<Room> list))
+            {
+                this._logger.LogInformation("RoomsController-GetRooms(hotelId) cache hit");
+            }
+            else
+            {
+                list = await this._context.Rooms.Include(h => h.Hotel).Where(h => h.Hotel.Id == hotelId).ToListAsync(token);
+
+                this._logger.LogInformation("RoomsController-GetRooms(hotelId) DB hit");
+
+                var cacheEntryOptions = new MemoryCacheEntryOptions().SetSlidingExpiration(TimeSpan.FromSeconds(40));
+
+                this._memoryCache.Set(key, list, cacheEntryOptions);
+            }
 
             var data = list.Select(r => this._mapper.Map<RoomResource>(r));
             return Ok(data);
         }
         [HttpGet("{id}")]
         //[ResponseCache(VaryByQueryKeys = new[] { "id" }, Duration = 30)]
-        public async Task<ActionResult<RoomResource>> Get(int hotelId, int id)
+        public async Task<ActionResult<RoomResource>> Get(int hotelId, int id, CancellationToken token)
         {
             if (id < 0)
             {
                 throw new ArgumentException("Negative Room id exception");
             }
 
-            var entity = await this._context.Rooms.FindAsync(id);
+            var hotel = await this._context.Hotels
+                    .Include(i => i.Rooms)
+                    .FirstOrDefaultAsync(i => i.Id == hotelId, token);
+
+            if (hotel == null) return this.NotFound();
+
+            var entity = await this._context.Rooms.FindAsync(new object[] { id }, token);
             if (entity == null)
             {
                 return this.NotFound();
@@ -85,9 +111,20 @@ namespace HotelApp.API.Controllers
         }
 
         [HttpPut("{id}")]
-        public async Task<IActionResult> Put(int hotelId, int id, UpdateRoomResource model)
+        public async Task<IActionResult> Put(int hotelId, int id, UpdateRoomResource model, CancellationToken token)
         {
-            var entity = await this._context.Rooms.FindAsync(id);
+            if (id < 0)
+            {
+                throw new ArgumentException("Negative Room id exception");
+            }
+
+            var hotel = await this._context.Hotels
+                    .Include(i => i.Rooms)
+                    .FirstOrDefaultAsync(i => i.Id == hotelId, token);
+
+            if (hotel == null) return this.NotFound();
+
+            var entity = await this._context.Rooms.FindAsync(new object[] { id }, token);
             if (entity == null)
             {
                 return this.NotFound();
@@ -97,22 +134,31 @@ namespace HotelApp.API.Controllers
             entity.ChildrenNumber = model.ChildrenNumber;
             entity.RoomName = model.RoomName;
             this._context.Rooms.Update(entity);
-            await this._context.SaveChangesAsync();
+            await this._context.SaveChangesAsync(token);
 
             return this.NoContent();
         }
 
         [HttpDelete("{id}")]
-        public async Task<ActionResult<Room>> Delete(int hotelId, int id)
+        public async Task<ActionResult<Room>> Delete(int hotelId, int id, CancellationToken token)
         {
-            var room = await this._context.Rooms.FindAsync(id);
+            var hotel = await this._context.Hotels
+                    .Include(i => i.Rooms)
+                    .FirstOrDefaultAsync(i => i.Id == hotelId, token);
+
+            if (hotel == null) return this.NotFound();
+
+            var room = await this._context.Rooms.FindAsync(new object[] { id }, token);
             if (room == null)
             {
                 return this.NotFound();
             }
 
             this._context.Rooms.Remove(room);
-            await this._context.SaveChangesAsync();
+            await this._context.SaveChangesAsync(token);
+
+            //inmemory cache invalidation
+            this._memoryCache.Remove($"_HOTEL_ROOMS_{hotelId}");
 
             return room;
         }
